+++ src/WordList/WordListInterface.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Stadly\PasswordPolice\WordList;
 
-use Stadly\PasswordPolice\Rule\TestException;
+use RuntimeException;
 
 /**
  * Interface that must be implemented by all word lists.
@@ -16,7 +16,7 @@ interface WordListInterface
      *
      * @param string $word Word to check.
      * @return bool Whether the word is present in the word list.
-     * @throws TestException If an error occurred while checking the word list.
+     * @throws RuntimeException If an error occurred while checking the word list.
      */
     public function contains(string $word): bool;
 }